Convert to class definition and add touch start/stop tracking
It turns out the combination of manually setting the coordinates to
create a looping effect and using the pagination feature of ScrollView
was causing the component to jitter.  By tracking certain events, we can
simulate the pagination behavior, while also setting up the component to
handle automatic periodic scrolling.

Solution uses PanHandler to fire an event when the user starts dragging the
carousel and when they stop dragging the carousel.

Had to convert the component to a class to take advantge of state -
specifically, PanHandler was having trouble using the current version of
the `interval` variable in its setTimeout callback.